requirements.txt
Flask>=2.0
Flask-Login
Flask-Mail
reportlab
werkzeug
pytesseract
pandas
openpyxl
python-dotenv
Pillow
